requirements.txt
pika===1.3.2
pyyaml===6.0.1